requirements.txt
pymongo >=4.7.1, <5.0.0
bson==0.5.10
python-dotenv==1.0.1
boto3>=1.35.14, <1.36